Switch to Comunica + Fuseki
For integration testing I needed a in-memory store.

Comunica is one of the few available. The returned
stream is slightly different from `sparql-http-client`,
and since we're using it for testing, we might as well
use Comunica also for the actual code.

Comunica and GraphDB were'nt playing nicely with each
other though, so replaced GraphDB with Fuseki.

Seems to be working ok so far